Implement Neutron IPv6 Prefix Delegation
This patch implements IPv6 Prefix Delegation for automatic allocation
of subnet CIDRs in Neutron. This is achieved using a new PrefixDelegation
class which provides an interface to a DHCPv6 client which handles PD
messages to and from a PD Server administered by the OpenStack deployer.

To enable this feature, subnets have been given a new property,
ipv6_pd_enabled. This property is hidden and does not affect the Neutron API.

WORKFLOW:

1. Admin sets default_ipv6_subnet_pool in neutron.conf to "prefix_delegation"
2. User creates a new IPv6 subnet without a CIDR or subnetpool ID
3. User creates an interface between this subnet and a router with an existing
external interface

NOTE: this patch requires minimum Dibbler client version of 1.0.1RC1
that is available at
https://github.com/tomaszmrugalski/dibbler/releases/tag/RELEASE1.0.1RC1

A PD Server reachable via the br-ex interface is needed to use this
feature. A dibbler-server configured to listen on br-ex is sufficient.

The pluggable PD client framework will be covered in a separate patch.

Co-Authored-By: Baodong (Robert) Li <baoli@cisco.com>

Change-Id: Ic0c6ed4dba74da94a75838178a1837f93d2d0885
Implements: blueprint ipv6-prefix-delegation
Depends-On: I77ccbffc5d04ccab705710b0c0ef144188eda711