Add an internal DHCPv6 client for Prefix Delegation
In order to use prefix delegation without the use of an external service
such as Dibbler this patch adds a small agent dedicated to handling
DHCPv6 Prefix Delegation which can run alongside the L3 Agent.

The agent can be started using the "neutron-dhcpv6-client" command and
must run with elevated privileges. The L3 Agent communicates with this
service via Unix sockets, using the PDManager we implement in the driver.

Co-Authored-By: Bradley Jones <jones.bradley@me.com>

Change-Id: Ic3849b4f0ad64804b642d00daadb9f265f2a7b62
Implements: blueprint neutron-ipv6-prefix-delegation-service